Att test

+++ contracts/Voting.sol
@@ -3,37 +3,52 @@ pragma solidity ^0.8.30;
 
 contract VotingPlatform {
     struct Candidate {
-        uint id;
+        uint256 id;
         string name;
-        uint voteCount;
+        uint256 voteCount;
     }
 
-    enum VotingState { NotStarted, Open, Closed }
+    enum VotingState {
+        NotStarted,
+        Open,
+        Closed
+    }
 
     struct Poll {
-        uint id;
+        uint256 id;
         address creator;
         string title;
         VotingState state;
-        uint candidateCount;
+        uint256 candidateCount;
     }
 
-    mapping(uint => Poll) public polls;
-    uint public pollCount;
+    mapping(uint256 => Poll) public polls;
+    uint256 public pollCount;
 
-    mapping(uint => mapping(uint => Candidate)) public candidates;
-    mapping(uint => mapping(address => bool)) public voters;
+    mapping(uint256 => mapping(uint256 => Candidate)) public candidates;
+    mapping(uint256 => mapping(address => bool)) public voters;
 
-    event PollCreated(uint indexed pollId, address indexed creator, string title);
-    event Voted(uint indexed pollId, uint indexed candidateId, address indexed voter);
-    event VotingStateChanged(uint indexed pollId, VotingState newState);
+    event PollCreated(
+        uint256 indexed pollId,
+        address indexed creator,
+        string title
+    );
+    event Voted(
+        uint256 indexed pollId,
+        uint256 indexed candidateId,
+        address indexed voter
+    );
+    event VotingStateChanged(uint256 indexed pollId, VotingState newState);
 
-    modifier onlyPollCreator(uint _pollId) {
-        require(polls[_pollId].creator == msg.sender, "Apenas o criador da votacao pode chamar esta funcao.");
+    modifier onlyPollCreator(uint256 _pollId) {
+        require(
+            polls[_pollId].creator == msg.sender,
+            "Apenas o criador da votacao pode chamar esta funcao."
+        );
         _;
     }
 
-    modifier pollExists(uint _pollId) {
+    modifier pollExists(uint256 _pollId) {
         require(_pollId > 0 && _pollId <= pollCount, "Votacao inexistente.");
         _;
     }
@@ -50,38 +65,75 @@ contract VotingPlatform {
         emit PollCreated(pollCount, msg.sender, _title);
     }
 
-    function addCandidate(uint _pollId, string memory _name) public pollExists(_pollId) onlyPollCreator(_pollId) {
+    function addCandidate(uint256 _pollId, string memory _name)
+        public
+        pollExists(_pollId)
+        onlyPollCreator(_pollId)
+    {
         Poll storage currentPoll = polls[_pollId];
-        require(currentPoll.state == VotingState.NotStarted, "A votacao ja comecou ou terminou.");
-        
+        require(
+            currentPoll.state == VotingState.NotStarted,
+            "A votacao ja comecou ou terminou."
+        );
+
         currentPoll.candidateCount++;
-        uint newCandidateId = currentPoll.candidateCount;
-        candidates[_pollId][newCandidateId] = Candidate(newCandidateId, _name, 0);
+        uint256 newCandidateId = currentPoll.candidateCount;
+        candidates[_pollId][newCandidateId] = Candidate(
+            newCandidateId,
+            _name,
+            0
+        );
     }
 
-    function startVoting(uint _pollId) public pollExists(_pollId) onlyPollCreator(_pollId) {
+    function startVoting(uint256 _pollId)
+        public
+        pollExists(_pollId)
+        onlyPollCreator(_pollId)
+    {
         Poll storage currentPoll = polls[_pollId];
-        require(currentPoll.state == VotingState.NotStarted, "A votacao ja foi iniciada.");
-        require(currentPoll.candidateCount > 0, "A votacao deve ter ao menos um candidato.");
-        
+        require(
+            currentPoll.state == VotingState.NotStarted,
+            "A votacao ja foi iniciada."
+        );
+        require(
+            currentPoll.candidateCount > 0,
+            "A votacao deve ter ao menos um candidato."
+        );
+
         currentPoll.state = VotingState.Open;
         emit VotingStateChanged(_pollId, VotingState.Open);
     }
 
-    function endVoting(uint _pollId) public pollExists(_pollId) onlyPollCreator(_pollId) {
+    function endVoting(uint256 _pollId)
+        public
+        pollExists(_pollId)
+        onlyPollCreator(_pollId)
+    {
         Poll storage currentPoll = polls[_pollId];
-        require(currentPoll.state == VotingState.Open, "A votacao nao esta aberta para ser encerrada.");
-        
+        require(
+            currentPoll.state == VotingState.Open,
+            "A votacao nao esta aberta para ser encerrada."
+        );
+
         currentPoll.state = VotingState.Closed;
         emit VotingStateChanged(_pollId, VotingState.Closed);
     }
 
-    function vote(uint _pollId, uint _candidateId) public pollExists(_pollId) {
+    function vote(uint256 _pollId, uint256 _candidateId)
+        public
+        pollExists(_pollId)
+    {
         Poll storage currentPoll = polls[_pollId];
 
-        require(currentPoll.state == VotingState.Open, "Esta votacao nao esta aberta.");
+        require(
+            currentPoll.state == VotingState.Open,
+            "Esta votacao nao esta aberta."
+        );
         require(!voters[_pollId][msg.sender], "Voce ja votou nesta votacao.");
-        require(_candidateId > 0 && _candidateId <= currentPoll.candidateCount, "Candidato invalido.");
+        require(
+            _candidateId > 0 && _candidateId <= currentPoll.candidateCount,
+            "Candidato invalido."
+        );
 
         candidates[_pollId][_candidateId].voteCount++;
         voters[_pollId][msg.sender] = true;
@@ -89,16 +141,26 @@ contract VotingPlatform {
         emit Voted(_pollId, _candidateId, msg.sender);
     }
 
-    function getAllCandidates(uint _pollId) public view pollExists(_pollId) returns (Candidate[] memory) {
-        uint count = polls[_pollId].candidateCount;
+    function getAllCandidates(uint256 _pollId)
+        public
+        view
+        pollExists(_pollId)
+        returns (Candidate[] memory)
+    {
+        uint256 count = polls[_pollId].candidateCount;
         Candidate[] memory allCandidates = new Candidate[](count);
-        for (uint i = 1; i <= count; i++) {
+        for (uint256 i = 1; i <= count; i++) {
             allCandidates[i - 1] = candidates[_pollId][i];
         }
         return allCandidates;
     }
 
-    function getVoterStatus(uint _pollId, address _voter) public view pollExists(_pollId) returns (bool) {
+    function getVoterStatus(uint256 _pollId, address _voter)
+        public
+        view
+        pollExists(_pollId)
+        returns (bool)
+    {
         return voters[_pollId][_voter];
     }
 }
