+++ tests/Voting.t.sol
@@ -1,14 +1,13 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.30;
 
-import "remix_tests.sol"; 
-import "../contracts/Voting.sol"; 
+import "remix_tests.sol";
+import "../contracts/Voting.sol";
 
 contract VotingPlatformTest {
-
     VotingPlatform votingPlatform;
 
-    address creator = address(this); 
+    address creator = address(this);
     address voter1 = address(0x1);
     address voter2 = address(0x2);
 
@@ -19,11 +18,25 @@ contract VotingPlatformTest {
     /// Testa a criacao de uma votacao
     function testCreatePoll() public {
         votingPlatform.createPoll("Quem deve vencer o concurso?");
-        (uint id,, string memory title,, uint candidateCount) = votingPlatform.polls(1);
+        (
+            uint256 id,
+            ,
+            string memory title,
+            ,
+            uint256 candidateCount
+        ) = votingPlatform.polls(1);
 
         Assert.equal(id, 1, "ID da votacao deve ser 1");
-        Assert.equal(title, "Quem deve vencer o concurso?", "Titulo da votacao incorreto");
-        Assert.equal(candidateCount, 0, "Votacao recem criada deve ter 0 candidatos");
+        Assert.equal(
+            title,
+            "Quem deve vencer o concurso?",
+            "Titulo da votacao incorreto"
+        );
+        Assert.equal(
+            candidateCount,
+            0,
+            "Votacao recem criada deve ter 0 candidatos"
+        );
     }
 
     /// Testa a adicao de candidatos
@@ -32,8 +45,14 @@ contract VotingPlatformTest {
         votingPlatform.addCandidate(1, "Modelo A");
         votingPlatform.addCandidate(1, "Modelo B");
 
-        (, string memory name1, uint votes1) = votingPlatform.candidates(1, 1);
-        (, string memory name2, uint votes2) = votingPlatform.candidates(1, 2);
+        (, string memory name1, uint256 votes1) = votingPlatform.candidates(
+            1,
+            1
+        );
+        (, string memory name2, uint256 votes2) = votingPlatform.candidates(
+            1,
+            2
+        );
 
         Assert.equal(name1, "Modelo A", "Nome do candidato 1 incorreto");
         Assert.equal(name2, "Modelo B", "Nome do candidato 2 incorreto");
@@ -58,14 +77,16 @@ contract VotingPlatformTest {
 
         // Simula o voto do contrato atual
         votingPlatform.vote(1, 1);
-        (, , uint voteCount) = votingPlatform.candidates(1, 1);
+        (, , uint256 voteCount) = votingPlatform.candidates(1, 1);
         Assert.equal(voteCount, 1, "Candidato 1 deve ter 1 voto");
 
         // Simula outro usuario votando via call
-        (bool success,) = address(votingPlatform).call(abi.encodeWithSignature("vote(uint256,uint256)", 1, 2));
+        (bool success, ) = address(votingPlatform).call(
+            abi.encodeWithSignature("vote(uint256,uint256)", 1, 2)
+        );
         Assert.ok(success, "Outro usuario deve conseguir votar");
 
-        (, , uint voteCount2) = votingPlatform.candidates(1, 2);
+        (, , uint256 voteCount2) = votingPlatform.candidates(1, 2);
         Assert.equal(voteCount2, 1, "Candidato 2 deve ter 1 voto");
     }
 
@@ -82,13 +103,17 @@ contract VotingPlatformTest {
         } catch {}
     }
 
-    /// Testa encerramento da votação
+    /// Testa encerramento da votacao
     function testEndVoting() public {
         votingPlatform.createPoll("Finalizar");
         votingPlatform.addCandidate(1, "A");
         votingPlatform.startVoting(1);
         votingPlatform.endVoting(1);
         (, , , VotingPlatform.VotingState state, ) = votingPlatform.polls(1);
-        Assert.equal(uint(state), uint(VotingPlatform.VotingState.Closed), "Votacao deveria estar encerrada");
+        Assert.equal(
+            uint256(state),
+            uint256(VotingPlatform.VotingState.Closed),
+            "Votacao deveria estar encerrada"
+        );
     }
 }
